+++ api/include/opentelemetry/context/context.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <bits/stdc++.h>
-#include <atomic>
 #include <map>
 #include <vector>
 
@@ -12,18 +11,34 @@
 #include "opentelemetry/trace/key_value_iterable_view.h"
 #include "opentelemetry/trace/key_value_iterable.h"
 #include "opentelemetry/context/key_value_iterable_modifiable.h"
-//#include "opentelemetry/context/threadlocal_context.h"
 
 OPENTELEMETRY_BEGIN_NAMESPACE
 namespace context
 {
 
-  /*The context class provides a context identifier */
 
+  /*The Key class is used to obscure access from the 
+   * user to the context map. The identifier is used as a key 
+   * to the context map.
+   */ 
   class Key
   {
+
+    public:
+
+      /* GetIdentifier: returns the identifier */
+      Key* GetIdentifier() { return identifier_; }
+
+
+      /* Constructs a new Key with the passed in name */
+      Key(nostd::string_view key_name)
+      {
+        key_name_ = key_name;
+        identifier_ = this;
+      }
+
     private:
-        template<class T>
+      template<class T>
         friend class Context;
 
       nostd::string_view key_name_;
@@ -31,43 +46,44 @@ namespace context
       Key* identifier_;
 
 
-
-      Key(nostd::string_view key_name, int identifier)
-      {
-        key_name_   = key_name;
-        identifier_ = this;
-      }
-
-    public:
-
-      Key* GetIdentifier() { return identifier_; }
-
-
-      Key(nostd::string_view key_name)
-      {
-        key_name_ = key_name;
-        identifier_ = this;
-      }
+      /*
+         Key(nostd::string_view key_name, int identifier)
+         {
+         key_name_   = key_name;
+         identifier_ = this;
+         }
+       */
   };
 
-    template<class T>
+  /* The context class provides a context identifier */
+  template<class T>
     class Context
     {
 
       public:
 
-
+        /* Creates a context object with no key/value pairs */
         Context() = default;
-        
+
+        /* Contructor, creates a context object from a map
+         * of keys and identifiers
+         */
         Context(const T &attributes) { 
           KeyValueIterableModifiable<T> iterable{attributes};
           key_val_map = iterable; 
         }
-        
+
+        /* Contructor, creates a context object from a 
+         * KeyValueIterableModfiable object.
+         */
         Context(const KeyValueIterableModifiable<T> iterable) { 
           key_val_map = iterable; 
         }
-        
+
+        /* Accepts a new iterable and then returns a new
+         * context that contains both the original pairs 
+         * and the new pair.
+         */
         Context WriteValues(const T &attributes) noexcept
         {
 
@@ -75,7 +91,8 @@ namespace context
           iterable.Add(key_val_map.GetData());
           return Context(iterable.GetData());
         }
-
+        
+        /* Returns the value associated with the passed in key */
         common::AttributeValue GetValue(Key key) { 
           return key_val_map.Get(key.GetIdentifier());
         }
@@ -84,69 +101,5 @@ namespace context
         KeyValueIterableModifiable<T> key_val_map;
 
     };
-
-
-
-  template <class M>
-  class Token;
-
-  template<class M>
-    class RuntimeContext
-    {
-
-      public:
-
-        Context<M> GetCurrent();
-        Token<M> Attach(Context<M> context);
-        int Detach(Token<M> token);
-    };
-
-   
-  template <class M>
-   class ThreadLocalContext : public RuntimeContext<M>
-   {
-      public:
-      
-      static Context<M> GetCurrent(){
-        return GetInstance();
-      }
-
-      int Detach(Token<M> token){
-        GetInstance() = token.GetCtx();
-        return 0;
-      }
-
-      Token<M> Attach(Context<M> context){
-        Token<M> old_context = Token<M>(GetInstance());
-        GetInstance() = context;
-        return old_context;
-      }
-
-      private:
-        
-        static Context<M> &GetInstance(){
-          static Context<M> instance;
-          return instance;
-        }
-       
-   };
-
-
-  template <class M>
-  class Token
-  {
-
-    private:
-      friend class RuntimeContext<M>;
-      friend class ThreadLocalContext<M>;
-
-      Context<M> ctx_;
-
-      Token(Context<M> ctx) { ctx_ = ctx; }
-
-
-      Context<M> GetCtx() { return ctx_; }
-  };
-
 }  // namespace context
 OPENTELEMETRY_END_NAMESPACE
